# Conflicts:
#	meta-ui-web/package.json

+++ api-types/src/data/Address.ts
@@ -1,4 +1,4 @@
-import type {bigserial} from '@/typescripts'
+import type {i32} from '@/typescripts'
 
 export type AddrCode = '00' | '000' | '' | `${number}`
 
@@ -15,5 +15,5 @@ export interface IChinaAddressDistrict {
   /**
    * @default 0
    */
-  level: bigserial
+  level: i32
 }
